@@ -31,7 +31,7 @@
     "sass-loader": "^7.1.0",
     "style-loader": "^0.23.0",
     "url-loader": "^1.1.1",
-    "webpack": "^4.17.2",
+    "webpack": "^4.28.3",
     "webpack-cli": "^3.1.0"
   },
   "dependencies": {
